stop ignoring runtime options (fixes issue #13)
This commit adds support for all runtime options that can be specified
with flags. To avoid breaking existing code that relies on running as
root, the drop_privileges runtime option is disabled by default, but can
be enabled by passing --led-drop-privs to your flags array.

We were using the RGBMatrix constructor directly, which doesn't deal
with runtime options (rgb_matrix::RuntimeOptions struct). This was
causing problems due to the fact that the --led-slowdown-gpio flag is
treated as a runtime option by the hzeller backend. Now we're using the
CreateMatrixFromFlags function which parses our flags for us and then
passes the options to CreateMatrixFromOptions, which deals with runtime
options and regular options for us.